Relocate @LockMethod and @UnlockMethod to new annotations package. ------------- Created by MOE: http://code.google.com/p/moe-java MOE_MIGRATED_REVID=84989739
