+++ annotations/src/main/java/com/google/errorprone/annotations/concurrent/LockMethod.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.errorprone.bugpatterns.threadsafety.annotations;
+package com.google.errorprone.annotations.concurrent;
 
 import static java.lang.annotation.ElementType.METHOD;
 import static java.lang.annotation.RetentionPolicy.CLASS;
