+++ core/src/main/java/com/google/errorprone/bugpatterns/threadsafety/LockMethodChecker.java
@@ -24,7 +24,7 @@ import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableSet;
 import com.google.errorprone.BugPattern;
 import com.google.errorprone.VisitorState;
-import com.google.errorprone.bugpatterns.threadsafety.annotations.LockMethod;
+import com.google.errorprone.annotations.concurrent.LockMethod;
 import com.google.errorprone.util.ASTHelpers;
 
 import com.sun.source.tree.MethodTree;
