+++ core/src/test/java/com/google/errorprone/bugpatterns/threadsafety/GuardedByLockMethodTest.java
@@ -17,8 +17,8 @@
 package com.google.errorprone.bugpatterns.threadsafety;
 
 import com.google.errorprone.CompilationTestHelper;
-import com.google.errorprone.bugpatterns.threadsafety.annotations.LockMethod;
-import com.google.errorprone.bugpatterns.threadsafety.annotations.UnlockMethod;
+import com.google.errorprone.annotations.concurrent.LockMethod;
+import com.google.errorprone.annotations.concurrent.UnlockMethod;
 
 import org.junit.Before;
 import org.junit.Test;
@@ -43,8 +43,8 @@ public class GuardedByLockMethodTest {
             "threadsafety/Test.java",
             "package threadsafety;",
             "import javax.annotation.concurrent.GuardedBy;",
-            "import com.google.errorprone.bugpatterns.threadsafety.annotations.LockMethod;",
-            "import com.google.errorprone.bugpatterns.threadsafety.annotations.UnlockMethod;",
+            "import com.google.errorprone.annotations.concurrent.LockMethod;",
+            "import com.google.errorprone.annotations.concurrent.UnlockMethod;",
             "import java.util.concurrent.locks.Lock;",
             "class Test {",
             "  final Lock lock = null;",
