+++ core/src/test/java/com/google/errorprone/bugpatterns/threadsafety/HeldLockAnalyzerTest.java
@@ -221,8 +221,8 @@ public class HeldLockAnalyzerTest {
             "threadsafety/Test.java",
             "package threadsafety;",
             "import javax.annotation.concurrent.GuardedBy;",
-            "import com.google.errorprone.bugpatterns.threadsafety.annotations.LockMethod;",
-            "import com.google.errorprone.bugpatterns.threadsafety.annotations.UnlockMethod;",
+            "import com.google.errorprone.annotations.concurrent.LockMethod;",
+            "import com.google.errorprone.annotations.concurrent.UnlockMethod;",
             "import java.util.concurrent.locks.Lock;",
             "class Outer {",
             "  Lock lock;",
