+++ core/src/test/java/com/google/errorprone/bugpatterns/threadsafety/LockMethodCheckerTest.java
@@ -36,7 +36,7 @@ public class LockMethodCheckerTest {
             "threadsafety/Test.java",
             "package threadsafety;",
             "import javax.annotation.concurrent.GuardedBy;",
-            "import com.google.errorprone.bugpatterns.threadsafety.annotations.LockMethod;",
+            "import com.google.errorprone.annotations.concurrent.LockMethod;",
             "import java.util.concurrent.locks.Lock;",
             "class Test {",
             "  Lock lock1;",
@@ -58,7 +58,7 @@ public class LockMethodCheckerTest {
             "threadsafety/Test.java",
             "package threadsafety;",
             "import javax.annotation.concurrent.GuardedBy;",
-            "import com.google.errorprone.bugpatterns.threadsafety.annotations.LockMethod;",
+            "import com.google.errorprone.annotations.concurrent.LockMethod;",
             "import java.util.concurrent.locks.Lock;",
             "class Test {",
             "  Lock lock1;",
@@ -81,7 +81,7 @@ public class LockMethodCheckerTest {
         compilationHelper.fileManager().forSourceLines(
             "threadsafety/Test.java",
             "package threadsafety;",
-            "import com.google.errorprone.bugpatterns.threadsafety.annotations.LockMethod;",
+            "import com.google.errorprone.annotations.concurrent.LockMethod;",
             "import java.util.concurrent.locks.ReentrantReadWriteLock;",
             "class Test {",
             "  ReentrantReadWriteLock lock;",
@@ -104,7 +104,7 @@ public class LockMethodCheckerTest {
         compilationHelper.fileManager().forSourceLines(
             "threadsafety/Test.java",
             "package threadsafety;",
-            "import com.google.errorprone.bugpatterns.threadsafety.annotations.LockMethod;",
+            "import com.google.errorprone.annotations.concurrent.LockMethod;",
             "import com.google.common.util.concurrent.Monitor;",
             "class Test {",
             "  Monitor monitor;",
@@ -124,7 +124,7 @@ public class LockMethodCheckerTest {
             "threadsafety/Test.java",
             "package threadsafety;",
             "import javax.annotation.concurrent.GuardedBy;",
-            "import com.google.errorprone.bugpatterns.threadsafety.annotations.LockMethod;",
+            "import com.google.errorprone.annotations.concurrent.LockMethod;",
             "import java.util.concurrent.locks.Lock;",
             "class Test {",
             "  Lock lock1;",
@@ -142,7 +142,7 @@ public class LockMethodCheckerTest {
         compilationHelper.fileManager().forSourceLines(
             "threadsafety/Test.java",
             "package threadsafety;",
-            "import com.google.errorprone.bugpatterns.threadsafety.annotations.LockMethod;",
+            "import com.google.errorprone.annotations.concurrent.LockMethod;",
             "import java.util.concurrent.locks.ReentrantReadWriteLock;",
             "class Test {",
             "  ReentrantReadWriteLock lock;",
@@ -159,7 +159,7 @@ public class LockMethodCheckerTest {
         compilationHelper.fileManager().forSourceLines(
             "threadsafety/Test.java",
             "package threadsafety;",
-            "import com.google.errorprone.bugpatterns.threadsafety.annotations.LockMethod;",
+            "import com.google.errorprone.annotations.concurrent.LockMethod;",
             "import com.google.common.util.concurrent.Monitor;",
             "class Test {",
             "  Monitor monitor;",
@@ -176,7 +176,7 @@ public class LockMethodCheckerTest {
         compilationHelper.fileManager().forSourceLines(
             "threadsafety/Test.java",
             "package threadsafety;",
-            "import com.google.errorprone.bugpatterns.threadsafety.annotations.LockMethod;",
+            "import com.google.errorprone.annotations.concurrent.LockMethod;",
             "class Test {",
             "  // BUG: Diagnostic contains: Could not resolve lock expression.",
             "  @LockMethod(\"mu\") void m() {}",
