+++ core/src/test/java/com/google/errorprone/bugpatterns/threadsafety/UnlockMethodCheckerTest.java
@@ -36,7 +36,7 @@ public class UnlockMethodCheckerTest {
             "threadsafety/Test.java",
             "package threadsafety;",
             "import javax.annotation.concurrent.GuardedBy;",
-            "import com.google.errorprone.bugpatterns.threadsafety.annotations.UnlockMethod;",
+            "import com.google.errorprone.annotations.concurrent.UnlockMethod;",
             "import java.util.concurrent.locks.Lock;",
             "class Test {",
             "  Lock lock1;",
@@ -57,7 +57,7 @@ public class UnlockMethodCheckerTest {
             "threadsafety/Test.java",
             "package threadsafety;",
             "import javax.annotation.concurrent.GuardedBy;",
-            "import com.google.errorprone.bugpatterns.threadsafety.annotations.UnlockMethod;",
+            "import com.google.errorprone.annotations.concurrent.UnlockMethod;",
             "import java.util.concurrent.locks.Lock;",
             "class Test {",
             "  Lock lock1;",
@@ -80,7 +80,7 @@ public class UnlockMethodCheckerTest {
         compilationHelper.fileManager().forSourceLines(
             "threadsafety/Test.java",
             "package threadsafety;",
-            "import com.google.errorprone.bugpatterns.threadsafety.annotations.UnlockMethod;",
+            "import com.google.errorprone.annotations.concurrent.UnlockMethod;",
             "import java.util.concurrent.locks.ReentrantReadWriteLock;",
             "class Test {",
             "  ReentrantReadWriteLock lock;",
@@ -103,7 +103,7 @@ public class UnlockMethodCheckerTest {
         compilationHelper.fileManager().forSourceLines(
             "threadsafety/Test.java",
             "package threadsafety;",
-            "import com.google.errorprone.bugpatterns.threadsafety.annotations.UnlockMethod;",
+            "import com.google.errorprone.annotations.concurrent.UnlockMethod;",
             "import com.google.common.util.concurrent.Monitor;",
             "class Test {",
             "  Monitor monitor;",
@@ -123,7 +123,7 @@ public class UnlockMethodCheckerTest {
             "threadsafety/Test.java",
             "package threadsafety;",
             "import javax.annotation.concurrent.GuardedBy;",
-            "import com.google.errorprone.bugpatterns.threadsafety.annotations.UnlockMethod;",
+            "import com.google.errorprone.annotations.concurrent.UnlockMethod;",
             "import java.util.concurrent.locks.Lock;",
             "class Test {",
             "  Lock lock1;",
@@ -141,7 +141,7 @@ public class UnlockMethodCheckerTest {
         compilationHelper.fileManager().forSourceLines(
             "threadsafety/Test.java",
             "package threadsafety;",
-            "import com.google.errorprone.bugpatterns.threadsafety.annotations.UnlockMethod;",
+            "import com.google.errorprone.annotations.concurrent.UnlockMethod;",
             "import java.util.concurrent.locks.ReentrantReadWriteLock;",
             "class Test {",
             "  ReentrantReadWriteLock lock;",
@@ -158,7 +158,7 @@ public class UnlockMethodCheckerTest {
         compilationHelper.fileManager().forSourceLines(
             "threadsafety/Test.java",
             "package threadsafety;",
-            "import com.google.errorprone.bugpatterns.threadsafety.annotations.UnlockMethod;",
+            "import com.google.errorprone.annotations.concurrent.UnlockMethod;",
             "import com.google.common.util.concurrent.Monitor;",
             "class Test {",
             "  Monitor monitor;",
@@ -175,7 +175,7 @@ public class UnlockMethodCheckerTest {
         compilationHelper.fileManager().forSourceLines(
             "threadsafety/Test.java",
             "package threadsafety;",
-            "import com.google.errorprone.bugpatterns.threadsafety.annotations.UnlockMethod;",
+            "import com.google.errorprone.annotations.concurrent.UnlockMethod;",
             "class Test {",
             "  // BUG: Diagnostic contains: Could not resolve lock expression.",
             "  @UnlockMethod(\"mu\") void m() {}",
